refactor: classes splitted in different files
